+++ render/d3d12/d3d12shader.h
@@ -79,7 +79,7 @@ namespace Lightning
 			container::vector<D3D12_ROOT_PARAMETER> mRootParameters;
 			container::unordered_map<std::string, D3D12_SHADER_INPUT_BIND_DESC> mInputBindDescs;
 			container::unordered_map<std::size_t, D3D12_SHADER_BUFFER_DESC> mBufferDescs;
-			container::unordered_map<std::uint8_t, container::vector<D3D12RootBoundResource>> mRootBoundResources;
+			container::unordered_map<std::size_t, container::vector<D3D12RootBoundResource>> mRootBoundResources;
 			//each offset corresponds to mIntermediateBuffer
 			container::unordered_map<std::size_t, ConstantBufferInfo> mConstantBufferInfo;
 			D3D12_DESCRIPTOR_RANGE *mDescriptorRanges;
