+++ render/vertexbuffer.h
@@ -16,17 +16,17 @@ namespace Lightning
 			const VertexComponent& GetComponentInfo(size_t attributeIndex)override;
 			~VertexBuffer()override;
 			//get vertex attribute count associate with this vertex buffer
-			std::uint8_t GetComponentCount()override;
+			std::size_t GetComponentCount()override;
 			//get vertices count contained within this vertex buffer
 			std::uint32_t GetVertexCount()const override;
 			//get vertex size in bytes
 			std::uint32_t GetVertexSize()const override;
 			void SetBuffer(std::uint8_t* buffer, std::uint32_t bufferSize)override;
 		protected:
-			VertexBuffer(uint32_t bufferSize, const VertexComponent *components, std::uint8_t componentCount);
+			VertexBuffer(uint32_t bufferSize, const VertexComponent *components, std::size_t componentCount);
 			void CalculateVertexSize();
 			VertexComponent *mComponents;
-			std::uint8_t mComponentCount;
+			std::size_t mComponentCount;
 			std::uint32_t mVertexCount;
 			std::uint32_t mVertexSize;
 		};
